requirements.txt
notion-client>=2.0.0
python-dotenv>=1.0.0
httpx>=0.27.0
playwright>=1.40.0
